refactor: retype accounts in transaction

+++ src/account.cpp
@@ -3,9 +3,9 @@
 #include <iostream>
 #include <string>
 
-Account::Account(int user_id, std::string account_number, int balance) {
+Account::Account(int user_id, std::string account_number, double balance) {
     this->user_id = user_id;
     this->account_number = account_number;
     this->balance = balance;
 }
-void Account::setBalance(int balance) { this->balance = balance; }
+void Account::setBalance(double balance) { this->balance = balance; }
